requirements.txt
aioredis==1.3.1
amqp==5.1.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
DateTime==4.4
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.5
django-property-filter==1.1.1
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httptools==0.3.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.0
msgpack==1.0.3
packaging==21.3
prompt-toolkit==3.0.28
# psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==4.2.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.2
whitenoise==6.0.0
zope.interface==5.4.0
